Added new typy a,b,c quadratic/forms.py Ver.5

+++ quadratic/forms.py
@@ -6,12 +6,12 @@ class QuadraticForm(forms.Form):
     """
     Input variables of the quadratic equation
     """
-    a = forms.FloatField(label="коэффициент a", widget=forms.TextInput)
-    b = forms.FloatField(label="коэффициент b", widget=forms.TextInput)
-    c = forms.FloatField(label="коэффициент c", widget=forms.TextInput)
+    a = forms.IntegerField(label="коэффициент a", widget=forms.TextInput)
+    b = forms.IntegerField(label="коэффициент b", widget=forms.TextInput)
+    c = forms.IntegerField(label="коэффициент c", widget=forms.TextInput)
 
     def clean_a(self):
-        if int(self.cleaned_data['a']) == 0:
+        if self.cleaned_data['a'] == 0:
             raise forms.ValidationError('коэффициент при первом слагаемом уравнения не может быть равным нулю')
         else:
             return self.cleaned_data['a']
